Video utils

+++ numpy_base/array_manipulation.py
@@ -4,15 +4,15 @@ import numexpr as ne
 
 def split_array_in_subarrays(data, n_subarrays, axis=0):
     """Split an array in subarrays. Don't need to have the same size.
-    
+
     Args:
         data (np.array): An array.
         n_subarrays (int): The number of subarrays.
         axis (int): Axis of the operation.
-    
+
     Returns:
         list: A list with subarrays.
-    
+
     Examples:
         >>> data = np.array([(1,2,3,4,5),(2,3,4,5,6)], dtype='int')
         >>> array_list = split_array_in_subarrays(data, 2, 1)
@@ -28,14 +28,14 @@ def split_array_in_subarrays(data, n_subarrays, axis=0):
 
 def concatenate_arrays(array_list, axis=0):
     """Concatenate a list of arrays.
-    
+
     Args:
         array_list (list): A list of arrays.
         axis (int): Axis of the operation.
-    
+
     Returns:
         np.array: An array of concatenated values.
-    
+
     Examples:
         >>> data1 = np.array([(1,2,3),(2,3,4)], dtype='int')
         >>> data2 = np.array([(4,5),(5,6)], dtype='int')
@@ -49,16 +49,16 @@ def concatenate_arrays(array_list, axis=0):
 
 def one_hot_encoding_integer(y, num_classes=None):
     """Converts a class vector (integers) to binary class matrix.
-    
+
     `See the original source <https://github.com/fchollet/keras/blob/d956d19fccf6de6344c282218f1b027453785fa9/keras/utils/np_utils.py>`_
-    
+
     Args:
         y (int): An integer class from 0 to num_classes.
         num_classes (int): Total number of classes.
-    
+
     Returns:
         np.array: An array of one hot encoded classes.
-    
+
     Examples:
         >>> one_hot_encoding_integer(2, 3)
         array([0, 0, 1])
@@ -80,20 +80,20 @@ def one_hot_encoding_integer(y, num_classes=None):
 
 def binarize_array(data, threshold, lower, upper):
     """Binarize an array based on a threshold within lower and upper values.
-    
+
     Args:
         data (np.array): An array.
         threshold (int or float): Threshold for binarization.
         lower (int or float): Lower value.
         upper (int or float): Upper value.
-    
+
     Returns:
         np.array: A binarized array.
-    
+
     Examples:
         >>> data = np.array([(1,3,5),(2,4,6)], dtype='int')
         >>> binarize_array(data, 3, 1, 6)
         array([[1, 1, 6],
-               [1, 6, 6]])
+               [1, 6, 6]], dtype=int64)
     """
     return ne.evaluate("where(data > threshold, upper, lower)")
